+++ src/config.ts
@@ -6,44 +6,34 @@ import {
 } from '@jupiterone/integration-sdk-core';
 import { createAPIClient } from './client';
 
-/**
- * A type describing the configuration fields required to execute the
- * integration for a specific account in the data provider.
- *
- * When executing the integration in a development environment, these values may
- * be provided in a `.env` file with environment variables. For example:
- *
- * - `CLIENT_ID=123` becomes `instance.config.clientId = '123'`
- * - `CLIENT_SECRET=abc` becomes `instance.config.clientSecret = 'abc'`
- *
- * Environment variables are NOT used when the integration is executing in a
- * managed environment. For example, in JupiterOne, users configure
- * `instance.config` in a UI.
- */
 export const instanceConfigFields: IntegrationInstanceConfigFieldMap = {
-  clientId: {
+  hostname: {
     type: 'string',
   },
-  clientSecret: {
+  token: {
     type: 'string',
     mask: true,
   },
+  namespace: {
+    type: 'string',
+  },
 };
 
-/**
- * Properties provided by the `IntegrationInstance.config`. This reflects the
- * same properties defined by `instanceConfigFields`.
- */
 export interface IntegrationConfig extends IntegrationInstanceConfig {
   /**
-   * The provider API client ID used to authenticate requests.
+   * The provider hostname used on requests.
    */
-  clientId: string;
+  hostname: string;
 
   /**
-   * The provider API client secret used to authenticate requests.
+   * The provider token used to authenticate requests.
    */
-  clientSecret: string;
+  token: string;
+
+  /**
+   * The provider namespace used to authenticate requests for cloud deployments.
+   */
+  namespace?: string;
 }
 
 export async function validateInvocation(
@@ -51,9 +41,9 @@ export async function validateInvocation(
 ) {
   const { config } = context.instance;
 
-  if (!config.clientId || !config.clientSecret) {
+  if (!config.hostname || !config.token) {
     throw new IntegrationValidationError(
-      'Config requires all of {clientId, clientSecret}',
+      'Config requires all of {hostname, token}',
     );
   }