+++ src/steps/index.ts
@@ -1,6 +1,19 @@
+import { secretEnginesSteps } from './secret-engine';
 import { accountSteps } from './account';
-import { accessSteps } from './access';
+import { cubbyholeSteps } from './cubbyhole';
+import { kv1Steps } from './kv1';
+import { kv2Steps } from './kv2';
+import { userSteps } from './user';
+import { authMethodsSteps } from './auth-backend';
 
-const integrationSteps = [...accountSteps, ...accessSteps];
+const integrationSteps = [
+  ...accountSteps,
+  ...userSteps,
+  ...secretEnginesSteps,
+  ...authMethodsSteps,
+  ...kv1Steps,
+  ...kv2Steps,
+  ...cubbyholeSteps,
+];
 
 export { integrationSteps };