+++ test/recording.ts
@@ -12,18 +12,26 @@ export function setupProjectRecording(
 ): Recording {
   return setupRecording({
     ...input,
-    redactedRequestHeaders: ['Authorization'],
+    redactedRequestHeaders: [
+      'Authorization',
+      'x-vault-token',
+      'x-vault-namespace',
+    ],
     redactedResponseHeaders: ['set-cookie'],
-    mutateEntry: mutations.unzipGzippedRecordingEntry,
-    /*mutateEntry: (entry) => {
+    // mutateEntry: mutations.unzipGzippedRecordingEntry,
+    mutateEntry: (entry) => {
       redact(entry);
-    },*/
+    },
+    options: {
+      matchRequestsBy: {
+        url: false,
+      },
+    },
   });
 }
 
 // a more sophisticated redaction example below:
 
-/*
 function getRedactedOAuthResponse() {
   return {
     access_token: '[REDACTED]',
@@ -55,20 +63,28 @@ function redact(entry): void {
   const responseText = entry.response.content.text;
   const parsedResponseText = JSON.parse(responseText.replace(/\r?\n|\r/g, ''));
 
-  //now we can modify the returned object as desired
-  //in this example, if the return text is an array of objects that have the 'tenant' property...
-  if (parsedResponseText[0]?.tenant) {
-    for (let i = 0; i < parsedResponseText.length; i++) {
-      parsedResponseText[i].client_secret = '[REDACTED]';
-      parsedResponseText[i].jwt_configuration = '[REDACTED]';
-      parsedResponseText[i].signing_keys = '[REDACTED]';
-      parsedResponseText[i].encryption_key = '[REDACTED]';
-      parsedResponseText[i].addons = '[REDACTED]';
-      parsedResponseText[i].client_metadata = '[REDACTED]';
-      parsedResponseText[i].mobile = '[REDACTED]';
-      parsedResponseText[i].native_social_login = '[REDACTED]';
+  if (parsedResponseText) {
+    if (parsedResponseText.data) {
+      if (parsedResponseText.data.id) {
+        parsedResponseText.data.id = '[REDACTED]';
+      }
     }
   }
 
+  //now we can modify the returned object as desired
+  //in this example, if the return text is an array of objects that have the 'tenant' property...
+  // if (parsedResponseText[0]?.tenant) {
+  //   for (let i = 0; i < parsedResponseText.length; i++) {
+  //     parsedResponseText[i].client_secret = '[REDACTED]';
+  //     parsedResponseText[i].jwt_configuration = '[REDACTED]';
+  //     parsedResponseText[i].signing_keys = '[REDACTED]';
+  //     parsedResponseText[i].encryption_key = '[REDACTED]';
+  //     parsedResponseText[i].addons = '[REDACTED]';
+  //     parsedResponseText[i].client_metadata = '[REDACTED]';
+  //     parsedResponseText[i].mobile = '[REDACTED]';
+  //     parsedResponseText[i].native_social_login = '[REDACTED]';
+  //   }
+  // }
+
   entry.response.content.text = JSON.stringify(parsedResponseText);
-} */
+}